+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.28.1] - 2024-07-10
+
+### Fixed
+- Sending empty system content in `PromptTask`.
+
 ## [0.28.0] - 2024-07-09
 ### Added
 - `RagEngine` is an abstraction for implementing modular RAG pipelines.